refactor: account type enume location
