+++ src/main/java/org/hoongoin/interviewbank/account/infrastructure/repository/AccountRepository.java
@@ -3,7 +3,7 @@ package org.hoongoin.interviewbank.account.infrastructure.repository;
 import java.util.Optional;
 
 import org.hoongoin.interviewbank.account.infrastructure.entity.AccountEntity;
-import org.hoongoin.interviewbank.account.infrastructure.entity.AccountType;
+import org.hoongoin.interviewbank.account.application.entity.AccountType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<AccountEntity, Long> {
