+++ src/main/java/org/hoongoin/interviewbank/interview/application/entity/DevelopSecondaryJobCategory.java
@@ -1,0 +1,2 @@
+package org.hoongoin.interviewbank.interview.application.entity;public class DevelopSecondaryJobCategory {
+}
